requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
gunicorn==20.1.0
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.3.1
python-dotenv==0.18.0
qrcode==7.1
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.5.0
